base_webapp_project/requirements.txt
flask
passlib
flask-wtf
email_validator
python-dotenv
sqlalchemy